Nvidia Support

+++ lib/functions.sh
@@ -10,7 +10,7 @@ function settings()
 		DRIVER=AMD
 	elif [ $DRIVER_NVIDIA -gt 0 ]; then
 		DRIVER=NVIDIA
-	fi	
+	fi
 	VERSION=0.51
 	URL="https://mrminer.co/api"
 	MAC=$(sudo ifconfig | grep -ioE '([a-z0-9]{2}:){5}..' | sed 's/://g' | sha256sum)
@@ -25,7 +25,7 @@ function boot()
 {
 	sudo rm -rf /home/mrminer/.cache/sessions/*
 	sudo rm -rf /etc/udev/rules.d/70-persistent-net.rules
-#	curl -k -d api="$API" -d email="$EMAIL" -d stats='{"total_hash": "", "total_hash_dual": "", "gpu_hash": "", "gpu_hash_dual": "", " "core": "", "mem": "", "temp": "", "fan": "", "watt": ""}' $URL/stats  /dev/null 2>&1 &
+	sudo rm -rf /etc/X11/xorg.conf
 	backup_oc_table
 	backup_miner
 	settings
@@ -133,17 +133,21 @@ function backup_miner()
 # Backup OC Table
 function backup_oc_table()
 {
-	sudo rm -rf /var/tmp/pp_tables && sleep 0.5 && sync && sudo mkdir /var/tmp/pp_tables
-	x=0
-	while [ $x -le 14 ]; do
-	    if [ -e "/sys/class/drm/card$x/device/pp_table" ]
-	    then
-	        sudo mkdir /var/tmp/pp_tables/gpu$x
-	        sudo cp /sys/class/drm/card$x/device/pp_table /var/tmp/pp_tables/gpu$x/pp_table
-					sync
-	    fi
-	    x=$[x + 1]
-	done
+	if [ "$DRIVER" == "AMD" ]; then
+
+		sudo rm -rf /var/tmp/pp_tables && sleep 0.5 && sync && sudo mkdir /var/tmp/pp_tables
+		x=0
+		while [ $x -le 14 ]; do
+		    if [ -e "/sys/class/drm/card$x/device/pp_table" ]
+		    then
+		        sudo mkdir /var/tmp/pp_tables/gpu$x
+		        sudo cp /sys/class/drm/card$x/device/pp_table /var/tmp/pp_tables/gpu$x/pp_table
+				sync
+		    fi
+		    x=$[x + 1]
+		done
+
+	fi
 }
 
 # Text Color
